Use escaped ASCII for unicode filename on python2
__repr__ with unicode is problematic in python2.

Interpolating a unicode filename into an ASCII string as previously
done  resulted in a UnicodeEncodeError.  We cannot interpolate a
unicode filename into a unicode string as the __repr__ function
must return a byte string in python2.  Since __repr__ does not
absolutely have to return a byte string that can always be used to
recreate the object, we just use REPR on the unicode filename under
python2, which results in escaped ASCII.

We cannot simply use repr(filename) for all cases, since it does change
the output in python3, changes something that isn't broken, and breaks
two existing tests anyway.

In python3 the issue doesn't arise since __repr__ and __str__ return
unicode instead of byte strings.

Closes-Issue:  #36